Time: 0 ms (100%), Space: 7.8 MB (94.68%) - LeetHub

+++ 0278-first-bad-version/0278-first-bad-version.cpp
@@ -6,7 +6,7 @@ public:
     int firstBadVersion(int n) {
         int low=1,high=n;
         while(low<=high){
-            long long mid=low+(high-low)/2;
+            int mid=low+(high-low)/2;
             if(isBadVersion(mid)){
                 high=mid-1;
             }
